# src/static/openapi.yaml
openapi: '3.0.1'
info:
  title: Inventory API
  description: Inventory API - Demo API
  version: 1.0.0
  termsOfService: 'https://api-inventory-single.playground.radix.equinor.com/terms.html'
  contact:
    name: 'Øyvind Rønne'
    url: 'https://github.com/equinor/api-strategy'
    email: oyron@equinor.com
servers:
  - url: 'https://api-dev.gateway.equinor.com/inventory-demo'
  - url: 'https://api-inventory-single.playground.radix.equinor.com/api'
tags:
  - name: inventory
    description: An inventory of books
paths:
  /books:
    get:
      tags:
        - inventory
      summary: Return all books
      security: 
        - oAuth2: [https://api-inventory-single.playground.radix.equinor.com/Inventory.books.ReadWrite]
      operationId: getBooks
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '401':
          description: Authorization information is missing
        '403':
          description: Authorization information is invalid
                  
    post:
        tags:
          - inventory
        summary: Add a new book to the inventory
        security: 
        - oAuth2: [https://api-inventory-single.playground.radix.equinor.com/Inventory.books.ReadWrite]
        operationId: addBook
        requestBody:
          description: bookData object that needs to be added to the inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookData'
        responses:
          '201':
            description: Created
            content:
              application/json:
                schema:
                  items:
                    $ref: '#/components/schemas/Book'
          '401':
            description: Authorization information is missing
          '403':
            description: Authorization information is invalid
                    
  /books/{id}:
    get:
      tags:
        - inventory
      summary: Get a specific book
      security: 
        - oAuth2: [https://api-inventory-single.playground.radix.equinor.com/Inventory.books.ReadWrite]      
      operationId: getBook
      parameters:
        - name: id
          in: path
          description: Book id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Book'
        '400':
          description: Bad request
        '401':
          description: Authorization information is missing
        '403':
          description: Authorization information is invalid
        '404':
          description: Not found

    put:
      tags:
        - inventory
      summary: Update existing book
      security: 
        - oAuth2: [https://api-inventory-single.playground.radix.equinor.com/Inventory.books.ReadWrite]
      operationId: updateBook
      parameters:
        - name: id
          in: path
          description: Book id
          schema:
            type: integer
          required: true
      requestBody:
          description: bookData object that needs to be added to the inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookData'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Book'
        '400':
          description: Bad request
        '401':
          description: Authorization information is missing
        '403':
          description: Authorization information is invalid
        '404':
          description: Not found

    delete:
      tags:
        - inventory
      summary: Delete existing book
      security: 
        - oAuth2: [https://api-inventory-single.playground.radix.equinor.com/Inventory.books.ReadWrite]
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          description: Book id
          schema:
            type: integer
          required: true
      responses:
        '204':
          description: No content
        '400':
          description: Bad request
        '401':
          description: Authorization information is missing
        '403':
          description: Authorization information is invalid
        '404':
          description: Not found

components:
  securitySchemes:
    oAuth2:
      type: oauth2
      description: This API uses OAuth 2 with the implicit grant flow. [More info](https://api.example.com/docs/auth)
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/3aa4a235-b6e2-48d5-9195-7fcf05b459b0/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/3aa4a235-b6e2-48d5-9195-7fcf05b459b0/oauth2/v2.0/token
          scopes:
            https://api-inventory-single.playground.radix.equinor.com/Inventory.books.ReadWrite: Read and write to book inventory
  schemas:
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
    BookData:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
  
externalDocs:
  description: API documentation
  url: 'https://github.com/oyron/inventory-api'